+++ pkg/cmd/cmd.go
@@ -86,8 +86,8 @@ func init() {
 				Category: "API RESOURCE",
 				Commands: []*cli.Command{
 					&imagesCreate,
-					&imagesRetrieve,
 					&imagesList,
+					&imagesGet,
 				},
 			},
 			{
@@ -95,12 +95,12 @@ func init() {
 				Category: "API RESOURCE",
 				Commands: []*cli.Command{
 					&instancesCreate,
-					&instancesRetrieve,
 					&instancesList,
-					&instancesPutInStandby,
-					&instancesRestoreFromStandby,
+					&instancesGet,
+					&instancesRestore,
 					&instancesLogs,
 					&instancesDelete,
+					&instancesStandby,
 				},
 			},
 			{
@@ -116,8 +116,17 @@ func init() {
 				Category: "API RESOURCE",
 				Commands: []*cli.Command{
 					&volumesCreate,
-					&volumesRetrieve,
 					&volumesList,
+					&volumesGet,
+				},
+			},
+			{
+				Name:     "ingresses",
+				Category: "API RESOURCE",
+				Commands: []*cli.Command{
+					&ingressesCreate,
+					&ingressesList,
+					&ingressesGet,
 				},
 			},
 			{